refactor(modelrouter): 移除 v1 代码，将 v2 提升为唯一版本
- 删除所有 v1 代码（16 个文件：anthropic.go, converter.go, handler.go, ir.go,
  openai_chat.go, openai_responses.go, protocol.go, resolver.go, store.go,
  stream_types.go, upstream.go 及对应测试文件）
- 将 v2/ 子目录下所有文件提升到 modelrouter/ 根目录
- 更新 3 个外部文件的 import 路径：modelrouter/v2 → modelrouter
  (worker/router, runtime/lifecycle/steps/model.go, model_test.go)
- 删除 DEPRECATED.md 和空的 v2/ 目录
