--- backend/internal/modelrouter/DEPRECATED.md
@@ -1,10 +1,0 @@
-# modelrouter (v1) — DEPRECATED
-
-This package is deprecated. Use `modelrouter/v2` instead.
-
-## Migration Guide
-- Change import: `modelrouter` → `modelrouter/v2`
-- `RegisterRoutes` signature is identical
-- `UpstreamConfig` fields are compatible
-- `DefaultStore()` and `DefaultProtocolForProvider()` work the same
-- New: `ProtocolGemini` protocol supported
