+++ backend/internal/modelrouter/adapter.go
