+++ backend/internal/modelrouter/protocol_openai_chat.go
