+++ backend/internal/modelrouter/protocol_openai_responses.go
