+++ backend/internal/modelrouter/protocol_openai_responses_test.go
