--- backend/internal/modelrouter/stream_types.go
@@ -1,4 +1,0 @@
-package modelrouter
-
-// Stream type constants are now defined in ir.go.
-// This file is intentionally left empty to avoid duplicate type definitions.
