+++ backend/internal/modelrouter/testdata/anthropic_stream_text.jsonl
