+++ backend/internal/modelrouter/testdata/anthropic_stream_tools.jsonl
