+++ backend/internal/modelrouter/testdata/anthropic_to_chat_response.json
