+++ backend/internal/modelrouter/testdata/anthropic_to_responses_request.json
