+++ backend/internal/modelrouter/testdata/chat_stream_tools.jsonl
