+++ backend/internal/modelrouter/testdata/gemini_to_chat_response.json
