+++ backend/internal/runtime/lifecycle/steps/model.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/insmtx/Leros/backend/internal/agent"
-	modelrouter "github.com/insmtx/Leros/backend/internal/modelrouter/v2"
+	modelrouter "github.com/insmtx/Leros/backend/internal/modelrouter"
 	"github.com/insmtx/Leros/backend/internal/worker/identity"
 )
